RAG_witn_LangChain/requirements.txt
python-dotenv 
langchain_community
langchain
langchain-openai
langchain_chroma
chromadb 
pypdf
pymupdf
pandas